feat: "관리자 페이지에서 MBTI 성향 생성작업 (Trait_id 값 안들어가는 문제 해결 완 ) "
